+++ App/Sources/plonk/main.swift
@@ -4,4 +4,5 @@ let app = NSApplication.shared
 let delegate = AppDelegate()
 app.delegate = delegate
 app.setActivationPolicy(.accessory)
+EditMenu.install(on: app)
 app.run()